修复 xxl_job_log 主键id类型为 bigint(20) ，  防止 int(11) 位溢出导致调度任务不能正常调度。

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/dao/XxlJobInfoDao.java
@@ -35,7 +35,7 @@ public interface XxlJobInfoDao {
 	
 	public int update(XxlJobInfo xxlJobInfo);
 	
-	public int delete(@Param("id") int id);
+	public int delete(@Param("id") long id);
 
 	public List<XxlJobInfo> getJobsByGroup(@Param("jobGroup") int jobGroup);
